Separated out the rest of the unit-tests from PublishedContentTests class.
The new fixtures should represent their intended feature.
The model classes are nested within the test fixture in order to keep them scoped and contained.
