+++ tests/Our.Umbraco.Ditto.Tests/ConversionHandlerTests.cs
@@ -1,13 +1,69 @@
-﻿using NUnit.Framework;
-
-using Our.Umbraco.Ditto.Tests.Mocks;
-using Our.Umbraco.Ditto.Tests.Models;
-
-namespace Our.Umbraco.Ditto.Tests
+﻿namespace Our.Umbraco.Ditto.Tests
 {
+    using global::Umbraco.Web;
+    using NUnit.Framework;
+    using Our.Umbraco.Ditto.Tests.Mocks;
+
     [TestFixture]
     public class ConversionHandlerTests
     {
+        #region Models
+
+        public class BaseCalculatedModel
+        {
+            [DittoIgnore]
+            public string AltText { get; set; }
+
+            [DittoOnConverting]
+            internal void BaseCalculatedModel_OnConverting(DittoConversionHandlerContext ctx)
+            {
+                AltText = ctx.Content.GetPropertyValue("prop1") + " " + ctx.Content.GetPropertyValue("prop2");
+            }
+        }
+
+        [DittoConversionHandler(typeof(CalculatedModelConversionHandler))]
+        public class CalculatedModel : BaseCalculatedModel
+        {
+            public string Name { get; set; }
+
+            [DittoIgnore]
+            public string AltText2 { get; set; }
+
+            [DittoOnConverted]
+            internal void CalculatedModel_OnConverted(DittoConversionHandlerContext ctx)
+            {
+                Name = "Test";
+            }
+        }
+
+        public class CalculatedModel2 : BaseCalculatedModel
+        {
+            [DittoIgnore]
+            public string AltText2 { get; set; }
+        }
+
+        #endregion
+
+        #region Conversion Handlers
+
+        public class CalculatedModelConversionHandler : DittoConversionHandler<CalculatedModel>
+        {
+            public override void OnConverted()
+            {
+                Model.AltText2 = Content.GetPropertyValue("prop1") + " " + Content.GetPropertyValue("prop2");
+            }
+        }
+
+        public class CalculatedModel2ConversionHandler : DittoConversionHandler<CalculatedModel2>
+        {
+            public override void OnConverted()
+            {
+                Model.AltText2 = Content.GetPropertyValue("prop1") + " " + Content.GetPropertyValue("prop2");
+            }
+        }
+
+        #endregion
+
         [Test]
         public void Can_Run_Conversion_Handlers()
         {
@@ -30,8 +86,8 @@ namespace Our.Umbraco.Ditto.Tests
 
             var model = content.As<CalculatedModel>();
 
-            Assert.That(model.AltText, Is.EqualTo("Test1 Test2"));
             Assert.That(model.Name, Is.EqualTo("Test"));
+            Assert.That(model.AltText, Is.EqualTo("Test1 Test2"));
             Assert.That(model.AltText2, Is.EqualTo("Test1 Test2"));
         }
